+++ src/components/cc-action-dispatcher/cc-action-dispatcher.types.d.ts
@@ -1,0 +1,5 @@
+interface Response {
+    numberOfPings: number;
+    numberOfPongs: number;
+    unresponsiveInstances: string[];
+}