+++ CHANGELOG.md
@@ -19,6 +19,8 @@ title: Changelog
 * `<cc-zone-input>`: make server markers not focusable.
 * `<cc-addon-admin>`: fix skeleton mode
 * `<cc-env-var-form>`: fix toggling to JSON mode while in skeleton state.
+* New component:
+  * `<cc-action-dispatcher>`
 
 ### For devs